Created apply_payments.php
All dependencies were changed too (I have ran a tool that searches for all callings of sl_eob_search.php and I changed them manually to apply_payments.php)
I've removed sl_eob_search.php also.
